+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 * Switch all your image fields to `:jsonb` types in migrations from `:text`.
 
+  ```
+  $ mix ecto.gen.migration switch_image_fields_to_jsonb
+  ```
+  Paste in:
+
+  ```
+  execute """
+  alter table images alter column image type jsonb using image::JSON
+  """
+
+  flush()
+
+  execute """
+  alter table users alter column avatar type jsonb using avatar::JSON
+  """
+  ```
+
+  Generate migrations for image fields and paste them into your migration
+  ```
+  Brando.Field.ImageField.generate_image_fields_migration()
+  ```
 
 
 ## v2.0.0-alpha.1-dev